Make JAX an optional dependency
pertpy is an analysis toolbox and most of it - Milo, Augur, the DGE
interface, Mixscape, Mixscale, enrichment, the perturbation spaces and 19
of 20 distance metrics - needs no autodiff runtime at all. Shipping
jaxlib to everyone is a large install for a minority of tools, and
`import pertpy` additionally called `config.update("jax_enable_x64", True)`
at module scope, silently changing the numerics of any other JAX code in
the user's session.

The JAX stack moves to a `jax` extra and every tool that needs it imports
it lazily. `import pertpy` no longer imports jax, numpyro, ott, flax or
optax, and no longer touches the x64 flag; that is now set from the
scCODA inference entry points instead.

- Drop `funsor`, which was a dependency but imported nowhere.
- Add a `jax` extra (ott-jax, numpyro, flax, optax). `tcoda` and `scgen`
  depend on it, so their installs are unchanged.
- Add `pertpy._jax.jax_import`, which turns a missing backend into an
  actionable hint naming the tool and the missing module. It inspects
  `ImportError.name` so unrelated failures propagate untouched.
- Gate Sccoda, Cinemaot and MLPClassifierSpace behind `tools.__getattr__`,
  joining the existing Tasccoda and Scgen pattern.
- Split the flax MLP out of `_discriminator_classifiers` into
  `_mlp_classifier`, so the sklearn-only LRClassifierSpace does not pull
  in a JAX install.
- Keep Distance and GuideAssignment eagerly importable; only the
  "wasserstein" metric and `assign_mixture_model` need the extra.
