+++ .pre-commit-config.yaml
@@ -37,7 +37,7 @@ repos:
       hooks:
           - id: mypy
             name: mypy
-            entry: uv run --extra dev --extra de --extra tcoda --extra test mypy src/pertpy tests
+            entry: uv run --extra dev --extra de --extra jax --extra tcoda --extra test mypy src/pertpy tests
             language: system
             types: [python]
             files: ^(src/pertpy|tests)/
