+++ src/pertpy/tools/__init__.py
@@ -1,8 +1,7 @@
 from importlib import import_module
 
+from pertpy._jax import jax_import
 from pertpy.tools._augur import Augur
-from pertpy.tools._cinemaot import Cinemaot
-from pertpy.tools._coda._sccoda import Sccoda
 from pertpy.tools._dialogue import Dialogue
 from pertpy.tools._distances._distance_tests import DistanceTest
 from pertpy.tools._distances._distances import Distance
@@ -12,10 +11,7 @@ from pertpy.tools._perturbation_efficacy._mixscale import Mixscale
 from pertpy.tools._perturbation_efficacy._mixscape import Mixscape
 from pertpy.tools._perturbation_space._clustering import ClusteringSpace
 from pertpy.tools._perturbation_space._comparison import PerturbationComparison
-from pertpy.tools._perturbation_space._discriminator_classifiers import (
-    LRClassifierSpace,
-    MLPClassifierSpace,
-)
+from pertpy.tools._perturbation_space._discriminator_classifiers import LRClassifierSpace
 from pertpy.tools._perturbation_space._simple import (
     CentroidSpace,
     DistanceSpace,
@@ -25,29 +21,41 @@ from pertpy.tools._perturbation_space._simple import (
     PseudobulkSpace,
 )
 
+_JAX_BACKED = {
+    "Cinemaot": ("pertpy.tools._cinemaot", "Cinemaot"),
+    "Sccoda": ("pertpy.tools._coda._sccoda", "Sccoda"),
+    "MLPClassifierSpace": ("pertpy.tools._perturbation_space._mlp_classifier", "MLPClassifierSpace"),
+}
+
 
 def __getattr__(name: str):
-    if name == "Tasccoda":
+    if name in _JAX_BACKED:
+        module_path, attr = _JAX_BACKED[name]
+        with jax_import(name):
+            return getattr(import_module(module_path), attr)
+    elif name == "Tasccoda":
         try:
             for extra in ["toytree", "ete4"]:
                 import_module(extra)
-            module = import_module("pertpy.tools._coda._tasccoda")
-            return module.Tasccoda
         except ImportError:
             raise ImportError(
                 "Extra dependencies required: toytree, ete4. Please install with: pip install toytree ete4"
             ) from None
+        with jax_import(name):
+            return import_module("pertpy.tools._coda._tasccoda").Tasccoda
     elif name in ["EdgeR", "PermutationTest", "PyDESeq2", "Statsmodels", "TTest", "WilcoxonTest"]:
         module = import_module("pertpy.tools._differential_gene_expression")
         return getattr(module, name)
     elif name == "Scgen":
-        try:
-            module = import_module("pertpy.tools._scgen")
-            return module.Scgen
-        except ImportError:
-            raise ImportError(
-                "Scgen requires scvi-tools to be installed. Please install with: pip install scvi-tools"
-            ) from None
+        with jax_import(name):
+            try:
+                return import_module("pertpy.tools._scgen").Scgen
+            except ImportError as error:
+                if (error.name or "").split(".")[0] != "scvi":
+                    raise
+                raise ImportError(
+                    "Scgen requires scvi-tools to be installed. Please install with: pip install 'pertpy[scgen]'"
+                ) from error
 
     raise AttributeError(f"module {__name__!r} has no attribute {name!r}")
