+++ src/pertpy/tools/_coda/_base_coda.py
@@ -39,9 +39,17 @@ if TYPE_CHECKING:
     from matplotlib.figure import Figure
     from seaborn.axisgrid import FacetGrid
 
-config.update("jax_enable_x64", True)
-
 RGBA = tuple[float, float, float, float]
+
+
+def _enable_x64() -> None:
+    """Switch JAX to double precision, which scCODA-type models need for stable inference.
+
+    JAX only exposes this as a process-wide flag, so it is set from the inference entry points
+    instead of at import time to avoid changing the numerics of unrelated JAX code on
+    ``import pertpy``.
+    """
+    config.update("jax_enable_x64", True)
 
 
 class CompositionalModel2(ABC):
@@ -118,6 +126,8 @@ class CompositionalModel2(ABC):
         """
         import arviz as az
         from numpyro.infer import Predictive
+
+        _enable_x64()
 
         mcmc_state = sample_adata.uns.get("scCODA_params", {}).get("mcmc", {})
         if "samples" not in mcmc_state:
@@ -385,6 +395,8 @@ class CompositionalModel2(ABC):
         Returns:
             Calls `self.__run_mcmc`
         """
+        _enable_x64()
+
         if isinstance(data, MuData):
             try:
                 sample_adata = data[modality_key]
@@ -447,6 +459,8 @@ class CompositionalModel2(ABC):
             >>> mdata = sccoda.prepare(mdata, formula="condition", reference_cell_type="Endocrine")
             >>> sccoda.run_hmc(mdata, num_warmup=100, num_samples=1000)
         """
+        _enable_x64()
+
         if isinstance(data, MuData):
             try:
                 sample_adata = data[modality_key]
